+++ TuEventoBackend/tu-evento/src/main/java/com/capysoft/tuevento/modules/category/domain/event/CategoryCreatedEvent.java
@@ -7,7 +7,7 @@ import lombok.Getter;
 @Builder
 public class CategoryCreatedEvent {
 
-    private final Long    categoryId;
+    private final Integer categoryId;
     private final String  name;
     private final boolean hasParent;
 }
